frontend: Extract TopBar from App
