+++ frontend/src/App.tsx
@@ -1,33 +1,16 @@
 import './i18n/config';
-import accountIcon from '@iconify/icons-mdi/account';
-import logoutIcon from '@iconify/icons-mdi/logout';
-import { Icon } from '@iconify/react';
 import { Box } from '@material-ui/core';
 import AppBar from '@material-ui/core/AppBar';
 import Container from '@material-ui/core/Container';
 import CssBaseline from '@material-ui/core/CssBaseline';
-import IconButton from '@material-ui/core/IconButton';
 import Link from '@material-ui/core/Link';
-import ListItemIcon from '@material-ui/core/ListItemIcon';
-import ListItemText from '@material-ui/core/ListItemText';
-import Menu from '@material-ui/core/Menu';
-import MenuItem from '@material-ui/core/MenuItem';
 import { makeStyles } from '@material-ui/core/styles';
 import Toolbar from '@material-ui/core/Toolbar';
 import { SnackbarProvider } from 'notistack';
 import React from 'react';
-import { useTranslation } from 'react-i18next';
 import { Provider } from 'react-redux';
-import {
-  BrowserRouter,
-  HashRouter,
-  Redirect,
-  Route,
-  RouteProps,
-  Switch,
-  useHistory,
-} from 'react-router-dom';
-import ThemeChangeButton from './components/App/ThemeChangeButton';
+import { BrowserRouter, HashRouter, Redirect, Route, RouteProps, Switch } from 'react-router-dom';
+import TopBar from './components/App/TopBar';
 import { ClusterTitle } from './components/cluster/Chooser';
 import ActionsNotifier from './components/common/ActionsNotifier';
 import AlertNotification from './components/common/AlertNotification';
@@ -35,10 +18,9 @@ import ReleaseNotes from './components/common/ReleaseNotes/ReleaseNotes';
 import Sidebar, { drawerWidth, NavigationTabs, useSidebarItem } from './components/Sidebar';
 import helpers from './helpers';
 import { useElectronI18n } from './i18n/electronI18n';
-import LocaleSelect from './i18n/LocaleSelect/LocaleSelect';
 import ThemeProviderNexti18n from './i18n/ThemeProviderNexti18n';
-import { getToken, setToken } from './lib/auth';
-import { useCluster, useClustersConf } from './lib/k8s';
+import { getToken } from './lib/auth';
+import { useClustersConf } from './lib/k8s';
 import { createRouteURL, getRoutePath, ROUTES } from './lib/router';
 import themes, { getThemeName, usePrefersColorScheme } from './lib/themes';
 import { getCluster } from './lib/util';
@@ -124,79 +106,6 @@ function RouteSwitcher() {
           )
         )}
     </Switch>
-  );
-}
-
-function TopBar() {
-  const appBarActions = useTypedSelector(state => state.ui.views.appBar.actions);
-  const [menuAnchorEl, setMenuAnchorEl] = React.useState(null);
-  const cluster = useCluster();
-  const history = useHistory();
-  const { t } = useTranslation('frequent');
-
-  function handleMenu(event: any) {
-    setMenuAnchorEl(event.currentTarget);
-  }
-
-  function handleClose() {
-    setMenuAnchorEl(null);
-  }
-
-  function hasToken() {
-    return !!cluster ? !!getToken(cluster) : false;
-  }
-
-  function logout() {
-    if (!!cluster) {
-      setToken(cluster, null);
-    }
-    setMenuAnchorEl(null);
-    history.push('/');
-  }
-
-  return (
-    <>
-      {
-        // @todo: Use a grid to compose the toolbar
-        Object.values(appBarActions).map((action, i) => (
-          <React.Fragment key={i}>{action()}</React.Fragment>
-        ))
-      }
-      <LocaleSelect />
-      <ThemeChangeButton />
-      <IconButton
-        aria-label={t('User menu')}
-        aria-controls="customized-menu"
-        aria-haspopup="true"
-        onClick={handleMenu}
-        color="inherit"
-      >
-        <Icon icon={accountIcon} />
-      </IconButton>
-
-      <span id="customized-menu">
-        <Menu
-          anchorEl={menuAnchorEl}
-          open={!!menuAnchorEl}
-          onClose={handleClose}
-          anchorOrigin={{
-            vertical: 'top',
-            horizontal: 'right',
-          }}
-          transformOrigin={{
-            vertical: 'top',
-            horizontal: 'right',
-          }}
-        >
-          <MenuItem component="a" onClick={logout} disabled={!hasToken()} dense>
-            <ListItemIcon>
-              <Icon icon={logoutIcon} />
-            </ListItemIcon>
-            <ListItemText primary="Log out" secondary={hasToken() ? null : '(No token set up)'} />
-          </MenuItem>
-        </Menu>
-      </span>
-    </>
   );
 }
